requirements.txt
fastapi[standard]
SQLAlchemy
pyjwt
pwdlib[argon2]
cryptography
starlette
psycopg2-binary
